cql3: expr: print only column name when failing to resolve column
resolve_column() is part of the prepare stage, and tries to
resolve a column name in a query against the table's columns.

If it fails, it prints the containing binary_expression as
context. However, that's unnecessary - the unresolved
column name is sufficient context. So print that.

The motivation is to unify preparation of binary_operator
left-hand-side and right-hand-side - prepare_expression()
doesn't have the extra parameter and it wouldn't make sense
to add it, as expressions might not be children of binary_operators.