tools: toolchain: build BOLT in the frozen toolchain
BOLT has been merged into mainline LLVM but it hasn't been packaged for Fedora
yet, so to use it we have to build it from source.
In addition, there is a critical concurrency bug in BOLT's intrumentation
runtime. Until it's fixed, we have to apply a patch to the source.