cql3: expr: use recursive descent when preparing subscripts
When encountering a subscript as the left-hand-side of a binary operator,
we assume the subscripted value is a column and process it directly.

As a step towards de-specializing the left-hand-side of binary operators,
use recursive descent into prepare_binop_lhs() instead. This requires
generating a column_specification for arbitrary expressions, so we
add a column_specification_of() function for that. Currently it will
return a good representation for columns (the only input allowed by
the grammar) and a bad representation (the text representation of the
expression) for other expressions. We'll have to improve that when we
relax the grammar.