compaction: Add test where compaction doesn't keep up with flush rate.
The test simulates a situation where N threads issue flushes to N
tables. Half are issuing lots of small flushes and the other half
issues a few large ones. This can lead to a situation where lots of
small sstables accumulate on disk, and, if compaction never has a chance
to keep up, resources can be exhausted.