mutation_cleaner: Fix accounting of space freed during version merging
Fixes the following assert failure:

  database_test: replica/memtable.cc:548: replica::flush_memory_accounter::~flush_memory_accounter(): Assertion `_mt._flushed_memory <= _mt.occupancy().used_space()' failed.

The assert complains that more memory was flushed than memtable
occupies. Version mering can reduce the amount of memory used by
memtable while it is being flushed. The problem is supposed to be
prevented by having the cleaner subtract from _flushed_memory when it
merges versions. But it was used total_space() instead of
used_space(), so it may subtract less than was actually freed.

Fixes #10780