Submodules: allow "Scylla only" fork
When scylla is forked, the modules are referenced in such a way that
forces forking all of the corresponding submodules too.
This commit changes this by directing the submodules at scylladb repos
instead of the sepcific user.

Tests:
Cloning scylla with ssh and https protocols and making sure the
submodules are checked out correctly (from scylladb).

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>