table: perform_offstrategy_compaction: move off-strategy logic to compaction_manager
compaction_manager needs to decide about running off-strategy
compaction or not based on the maintenance_set, not partly
in table::trigger_offstrategy_compaction and part in
the compaction_manager layer as it is done today.

So move the logic down to performa_offstrategy
that now returns future<bool> to return true
iff it performed offstrategy compaction.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>